Reorganizes components folder and updates imports for restaurant hook and new component file locations
