+++ app/(modal)/foodItem.tsx
@@ -12,9 +12,8 @@ import { useNavigation } from "expo-router";
 const foodItem = () => {
   const navigation = useNavigation();
 
-  const { item, deliveryTime } = useFoodItem();
-
-  const { addProduct, setTime } = useCart();
+  const { item } = useFoodItem();
+  const { addProduct } = useCart();
 
   const addToCart = (item: FoodItem) => {
     addProduct(item);
@@ -49,7 +48,6 @@ const foodItem = () => {
             style={tw`h-12 px-4 rounded-md bg-green-700 items-center justify-center`}
             onPress={() => {
               addToCart(item!);
-              setTime(deliveryTime);
             }}
           >
             <Text style={tw`text-white font-semibold`}>
