+++ components/Parallax/ParallaxHeader.tsx
@@ -20,7 +20,7 @@ import tw from "twrnc";
 import { Restaurant } from "../../constants/types";
 
 import ParallaxScrollView from "./ParallaxScrollView";
-import FoodItemCard from "../FoodItemCard";
+import FoodItemCard from "../Cards/FoodItemCard";
 import StickyHeader from "./StickyHeader";
 
 interface ParallaxHeaderProps {
